requirements.txt
asgiref==3.4.1
backoff==1.10.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
croniter==1.0.13
Deprecated==1.2.12
docker==5.0.0
fastapi==0.63.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
h11==0.12.0
idna==2.10
motor==2.3.1
odmantic==0.3.4
opentelemetry-api==1.4.1
opentelemetry-exporter-otlp==1.4.1
opentelemetry-exporter-otlp-proto-grpc==1.4.1
opentelemetry-instrumentation==0.23b2
opentelemetry-instrumentation-asgi==0.23b2
opentelemetry-instrumentation-fastapi==0.23b2
opentelemetry-instrumentation-pymongo==0.23b2
opentelemetry-instrumentation-redis==0.23b2
opentelemetry-proto==1.4.1
opentelemetry-sdk==1.4.1
opentelemetry-semantic-conventions==0.23b2
opentelemetry-util-http==0.23b2
protobuf==3.17.3
pydantic==1.8.1
pymongo==3.11.4
python-dateutil==2.8.1
redis==3.5.3
requests==2.25.1
rq==1.8.0
rq-scheduler==0.11.0
six==1.16.0
starlette==0.13.6
supervisor==4.2.2
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
websocket-client==0.59.0
wrapt==1.12.1
